+++ ballcat-common/ballcat-common-idempotent/src/main/java/com/hccake/ballcat/common/idempotent/IdempotentAspect.java
@@ -3,8 +3,8 @@ package com.hccake.ballcat.common.idempotent;
 import cn.hutool.core.lang.Assert;
 import com.hccake.ballcat.common.idempotent.annotation.Idempotent;
 import com.hccake.ballcat.common.idempotent.exception.IdempotentException;
-import com.hccake.ballcat.common.idempotent.key.IdempotentKeyStore;
-import com.hccake.ballcat.common.idempotent.key.KeyGenerator;
+import com.hccake.ballcat.common.idempotent.key.generator.KeyGenerator;
+import com.hccake.ballcat.common.idempotent.key.store.IdempotentKeyStore;
 import com.hccake.ballcat.common.model.result.BaseResultCode;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
