+++ ballcat-common/ballcat-common-idempotent/src/main/java/com/hccake/ballcat/common/idempotent/key/generator/KeyGenerator.java
@@ -1,4 +1,4 @@
-package com.hccake.ballcat.common.idempotent.key;
+package com.hccake.ballcat.common.idempotent.key.generator;
 
 import com.hccake.ballcat.common.idempotent.annotation.Idempotent;
 import org.aspectj.lang.JoinPoint;
