+++ ballcat-common/ballcat-common-idempotent/src/main/java/com/hccake/ballcat/common/idempotent/key/store/RedisIdempotentKeyStore.java
@@ -1,4 +1,4 @@
-package com.hccake.ballcat.common.idempotent.key;
+package com.hccake.ballcat.common.idempotent.key.store;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
