+++ ballcat-dependencies/pom.xml
@@ -200,7 +200,17 @@
 			</dependency>
 			<dependency>
 				<groupId>com.hccake</groupId>
+				<artifactId>ballcat-spring-boot-starter-idempotent</artifactId>
+				<version>${revision}</version>
+			</dependency>
+			<dependency>
+				<groupId>com.hccake</groupId>
 				<artifactId>ballcat-spring-boot-starter-i18n</artifactId>
+				<version>${revision}</version>
+			</dependency>
+			<dependency>
+				<groupId>com.hccake</groupId>
+				<artifactId>ballcat-spring-boot-starter-ip2region</artifactId>
 				<version>${revision}</version>
 			</dependency>
 			<dependency>
